+++ client/doublezero/src/command/status.rs
@@ -226,6 +226,10 @@ mod tests {
                     doublezero_serviceability::state::device::DeviceHealth::ReadyForUsers,
                 desired_status:
                     doublezero_serviceability::state::device::DeviceDesiredStatus::Activated,
+                unicast_users_count: 0,
+                multicast_users_count: 0,
+                max_unicast_users: 0,
+                max_multicast_users: 0,
             },
         );
 
@@ -255,6 +259,10 @@ mod tests {
                     doublezero_serviceability::state::device::DeviceHealth::ReadyForUsers,
                 desired_status:
                     doublezero_serviceability::state::device::DeviceDesiredStatus::Activated,
+                unicast_users_count: 0,
+                multicast_users_count: 0,
+                max_unicast_users: 0,
+                max_multicast_users: 0,
             },
         );
 
@@ -425,6 +433,10 @@ mod tests {
                     doublezero_serviceability::state::device::DeviceHealth::ReadyForUsers,
                 desired_status:
                     doublezero_serviceability::state::device::DeviceDesiredStatus::Activated,
+                unicast_users_count: 0,
+                multicast_users_count: 0,
+                max_unicast_users: 0,
+                max_multicast_users: 0,
             },
         );
 
@@ -454,6 +466,10 @@ mod tests {
                     doublezero_serviceability::state::device::DeviceHealth::ReadyForUsers,
                 desired_status:
                     doublezero_serviceability::state::device::DeviceDesiredStatus::Activated,
+                unicast_users_count: 0,
+                multicast_users_count: 0,
+                max_unicast_users: 0,
+                max_multicast_users: 0,
             },
         );
 
@@ -598,6 +614,10 @@ mod tests {
                 max_users: 128,
                 desired_status: DeviceDesiredStatus::Activated,
                 device_health: DeviceHealth::ReadyForUsers,
+                unicast_users_count: 0,
+                multicast_users_count: 0,
+                max_unicast_users: 0,
+                max_multicast_users: 0,
             },
         );
 
@@ -746,6 +766,10 @@ mod tests {
                 max_users: 128,
                 device_health: DeviceHealth::ReadyForUsers,
                 desired_status: DeviceDesiredStatus::Activated,
+                unicast_users_count: 0,
+                multicast_users_count: 0,
+                max_unicast_users: 0,
+                max_multicast_users: 0,
             },
         );