+++ smartcontract/programs/doublezero-serviceability/tests/device_update_location_test.rs
@@ -369,6 +369,8 @@ async fn device_update_location_test() {
             desired_status: None,
             resource_count: 0,
             reference_count: None,
+            max_unicast_users: None,
+            max_multicast_users: None,
         }),
         vec![
             AccountMeta::new(device_pubkey, false),
@@ -412,6 +414,8 @@ async fn device_update_location_test() {
             desired_status: None,
             resource_count: 0,
             reference_count: None,
+            max_unicast_users: None,
+            max_multicast_users: None,
         }),
         vec![
             AccountMeta::new(device_pubkey, false),