+++ smartcontract/sdk/rs/src/commands/device/sethealth.rs
@@ -77,6 +77,10 @@ mod tests {
             users_count: 0,
             device_health: DeviceHealth::ReadyForUsers,
             desired_status: DeviceDesiredStatus::Activated,
+            unicast_users_count: 0,
+            multicast_users_count: 0,
+            max_unicast_users: 0,
+            max_multicast_users: 0,
         };
 
         client
@@ -120,6 +124,8 @@ mod tests {
                     desired_status: None,
                     resource_count: 2,
                     reference_count: None,
+                    max_unicast_users: None,
+                    max_multicast_users: None,
                 })),
                 predicate::always(),
             )
@@ -141,6 +147,8 @@ mod tests {
             status: None,
             desired_status: None,
             reference_count: None,
+            max_unicast_users: None,
+            max_multicast_users: None,
         };
 
         let update_invalid = UpdateDeviceCommand {